fix(vault): Fix deposit LP calculation
Related to #1 found in the SCV audit done.

Currently, the following scenario incorrectly incentivises actors to
continuously deposit and withdraw to exploit the profit held in the
vault contract.
If Alice deposits into the vault ($10,000 uluna, and receives 10,000 LP
tokens in return) and the vault experiences profits up to $15,000 uluna,
then the next depositor (depositing $5,000 uluna) will receive 5,000 LP
tokens which are worth more than proportional to their deposit amount.

Currently, the second depositor would be able to withdraw $6,666
(greater than their $5,000 deposit) despite not being an early holder of
the LP token before the vault inflated in value.

This commit correctly patches this behavior so that Alice is rewarded
with the $5,000 profits, while Bob is only able to withdraw their
initial $5,000 deposit.