fix(frontend_contract): transfer cw20 assets from user to pool contract
Before, a user firing the `frontend_contract::Deposit` function would
error if the user was attempting to deposit with CW20 tokens. This is
because the contract would not transfer the cw20 token from the user to
the frontend contract before increasing the allowance on the frontend
contract to allow the pool to provide the liquidity.

This commit rectifies the issue by performing a TransferFrom message
from the user to the frontend contract before increasing the allowance
to allow the pool pair to transfer from the frontend_contract to the
pool.