fix(incentive): Fix weight function calculating weight too small
Previously the weight function would potentially calculate a multiplier
of less than one (as the computation is only 18 decimal places). This
causes issues as the remaining code (namely, closing positions) depends
on the behavior of the weight calculation to return a value that is at
least as great as the amount passed in.

This commit performs a hot patch by ensuring that the returned value by
the weight function is at least the input passed. Effectively, the
previous multiplier of 0.999999999999999998 is now rounded to 1.0.