feat(migrate): Check contract names when performing migrations
Previous we only checked contract versions. This means in theory we could replace the currently deploy 1.0.0 instance of a contract with any other contract instance provided the version matches

Now we are checking the name to ensure we haven't mistakenly or accidentally performed a migration to another contract. To do that sort of logic where we want a new contract name but the same contract version this check should be removed.

Attempted to test this but there is a problem testing it in unit tests. Gonna try again with integration later. In particular with stableswap 3pool tests I want to add anyways

Fixes Check contract names when performing migrations #175