refactor(pool-manager): remove redundant assert_operations check
The previous commit changes the check inside assert_operations to ensure
that the output of each swap acts as the input of the next swap.

Naturally, this ensures that each swap does not result in multiple
outputs. We therefore simplify the code by removing this check.

This has an added benefit of removing our `HashMap` usage.